Reduce flickering by drawing into a Pixmap first
Essentially, do double buffering.

Why does this even matter? Modern toolkits set the property
_NET_WM_USER_TIME on client windows. It's being updated when there is
user activity on that window. This happens, for example, when you press
and hold Enter in a VTE terminal. Since we redraw our window on
PropertyNotify events, this resulted in flickering.

This is an extreme case, of course. But double buffering is not hard to
implement, so why not do it?